+++ R/roxygen-functions.R
@@ -13,9 +13,8 @@ make_namespace_S3_entries <- function (accessors) {
 }
 
 
-make_exports <- function (properties, with_by = FALSE) {
-  fun_templates <- c('%s', '"%s<-"', 'set_%s')
-  if (with_by) fun_templates <- c(fun_templates, 'set_%s_by')
+make_exports <- function (properties, with_map = FALSE) {
+  fun_templates <- c('%s', '"%s<-"', 'set_%s', if (with_map) 'map_%s')
   funs <- c(outer(fun_templates, properties, sprintf))
 
   paste0('export(', funs ,')')