+++ R/zzz.R
@@ -51,7 +51,7 @@ NULL
 #' These functions are deprecated and will be removed in future versions of huxtable.
 #'
 #' To replace the 3 argument form of `set_xxx` functions, use
-#' \code{\link[=mapping-functions]{set_xxx_by}}.
+#' \code{\link[=mapping-functions]{map_xxx}}.
 #' @name huxtable-deprecated
 NULL