+++ man-roxygen/cell-property-usage.R
@@ -4,4 +4,4 @@
 #' <%= attr_name %>(ht)
 #' <%= attr_name %>(ht) <- value
 #' set_<%= attr_name %>(ht, row, col, value, byrow = FALSE)
-#' set_<%= attr_name %>_by(ht, row, col, fn)
+#' map_<%= attr_name %>(ht, row, col, fn)